Update README.md
Added a line to tell Mac users to clone the repo after they install the Brew dependencies, just to make things a little clearer. 

I changed some of the wording when cloning the repository. Capitalised the word GitHub.

In addition, adding a line to tell the user to cd into the nomacs directory once cloned.