Update Git revision tagging
Building Nomacs from the zip source archives
downloaded from Github fails because the `versionupdate.py` script
calls Git command, which assumes the working direction is a child of
a Git repo.

Move the revision tag logic to the CMake script. Check for the existence
of a `.git` directory to fix the issue.

Also, include the commit date and check whether there are modified files
in the repository. These could be useful when reporting and searching
for bugs. Assemble the result into a constant string for use.

Fixes https://github.com/nomacs/nomacs/issues/1034.