Include the revision string in about nomacs
The NOMACS_GIT_TAG was defined but not referenced anywhere. Add a new
line in the version info of the about splash screen to display the
updated revision string, which contains the commit hash, the commit
date, and whether the source has modifications that have not been
committed when building. This could be helpful for reporting bugs.