Change request URLs to use HTTPS
Nomacs used HTTP to request for things like version files for auto
updates. Because the version file contains URL, if the file gets
tempered, the user could end up downloading malicious files. Modify the
URLs used in the code so that requests use HTTPS by default, which
should be safer.