Fix incorrect crop tooltip when rotating (#1073)
Fixes https://github.com/nomacs/nomacs/issues/1071

The tooltip of the crop tool displays the center coordinates of the crop
rectangle when it is rotated, however, the displayed value changes when
we rotate around the center.

The value is incorrect because the center is not properly translated
before applying the rotation transform, which rotates around the
origin. This patch calculates the coordinates to be displayed
considering only the translation because rotation around the center
should not change its position.