Fix segfault in history dock when last tab is closed
If the last tab is closed or we switch to a tab with no image (settings)
we will crash due to receiving a null image from imageUpdatedSignal().
Passing of null seems to be expected behavior as it indicates to the
widget there is no longer any image present.

This adds the null check where needed. We must clear the history since
restoring history of a detached image, while possible seems incorrect
when "ask to save changes" prompt should take care of this.