Fix segfaults or hangs when sorting (#1095)
The current version will crash if certain network folders are sorted by
created time. This is because network shares do not typically support
created/birth time.

Since the birth time is unavailable, the comparison result is always
false (this is how Qt compares invalid values). With inversion/
descending, then it is always true and std:sort crashes or hangs. If we
want to fix this for now and add future sort features this has to
change.

Related: #953 which crashes for this very issue; the "coin flip" happens
to come up 0 or 1 for all images being sorted.

I have rewritten DkImageLoader::sort() and refactored. I removed the
compare operators since they are unsafe and the old sorting method; some
dead code for threaded sorting remains, maybe someone will resurrect it
(not theoretically a bad idea).

The new version will prevent this bug from ever occuring by reversing
the list after sorting if needed. It should be a little faster since the
invariants are pulled out before sorting.