Add Cmake cross-compilation target for Mingw/Windows (#1097)
Adds necessary cmake files to compile the Windows version on a Linux/
Unix host. Also includes a script to make a Windows portable package.

This is intended for MXE as the cross compiling environment, since its
the easiest method. However, it also works with Mys2/mingw target.

Also fix tabs in the affected files