Startup: fix non-maximized window at startup on GNOME
On some platforms (GNOME/X11) and potentially other X11 window managers,
the maximized state of the main window will not be restored.

Since Qt emulates showMaximized() on X11, it is likely this is not
limited to GNOME and could affect other WMs now and in the future.

This corrects the problem on GNOME and may also prevent other WMs from
breaking in the future.

I have also added a more complete "ensure the window is visible/active"
check as sendPostedEvents() is not quite enough to fully realize the
window.