Fix navigation loss after file deletion
Fixes the problem where deleting the last file in a directory with
looping off causes nav loss.

This was because getSkippedImage(1) would not take the correct path
because mImages was not updating. To avoid rescanning the directory (an
alternate solution), this patch simply removes the path from mImages
after deletion.

The behavior remains the same regardless of directory mode:
- if looping, deletion at the end goes back to the start
- otherwise, it always advances based on the current sort, potentially
to the next subdirectory if that is enabled.
- if we cannot advance for any reason, we attempt to load the new image
at the current position

Fixes #600