Fix incorrect positioning when sorted by metadata
Positioning can be incorrect after deletion or other changes to
directory when sorted by anything which has the same value for multiple
files. This is not uncommon, for example burst shots or batched
processed files often have the same ctime/mtime.

This happens because getSkippedImage() re-establishes the position in
the folder based on the sorted position of the missing item.

To sidestep the issue, sort files by filename whenever their metadata
matches. So there should never be a case when the position is ambiguous.

This handles the case when we delete files and lose our position, or the
files were modified externally.

Fixes #391