Rewrite DkFileNameConverter
The issue https://github.com/nomacs/nomacs/issues/1139
is caused by `DkFileNameConverter` not parsing the input pattern
correctly. For example, "<c:0>.<old>" is interpreted as "<c:0>" because
that the regex "<.*>" gets the whole string instead of two groups.
This rewirtes `DkFileNameConverter` to parse the pattern manually
instead of using regex.

The signature of the converter is also changed, so we can create one
converter and reuse it instead of creating one in each iteration.

Fixes https://github.com/nomacs/nomacs/issues/1139.