improve interactivity in recursive scanner
We add a progress dialog that will show progress then also improve on
logging so that we show timing information.

While we're here, we remove debugging code that was just confusing the
reader: getFoldersRecursive() is not actually designed to be called
recursively, as the QDirIterator class does that for us.