fix: get existing thumbnails when setting loader
DkViewPort reuses the DkImageLoader from the last tab. Because the
ImageLoader points to a directory when using "Open Directory",
the next image load from that directory does not count as a directory
change. This makes some signal not emitted, so the new ViewPort does not
receive any thumbnail.

Get the existing images from the loader when setting the loader to the
viewport.

Fixes https://github.com/nomacs/nomacs/issues/763