Fix thumbnail not loaded bug and refactor DkImageLoader (#1181)
* refactor: remove unused methods in DkImageLoader

Remove methods in DkImageLoader that are not referenced anywhere.

* refactor: remove unused keywords in DkImageLoader

The mKeywords and mIgnoreKeywords are not written, so they are always
empty. Remove the logic around them.

* refactor: reduce public methods of DkImageLoader

Change methods that are not used outside of DkImageLoader to protected to
reduce the public interface.

* refactor: remove getSavePath of DkImageLoader

Because mSaveDir is never written, the getSavePath() method will always
return mCurrentDir. We can remove the function.

* fix: get existing thumbnails when setting loader

DkViewPort reuses the DkImageLoader from the last tab. Because the
ImageLoader points to a directory when using "Open Directory",
the next image load from that directory does not count as a directory
change. This makes some signal not emitted, so the new ViewPort does not
receive any thumbnail.

Get the existing images from the loader when setting the loader to the
viewport.