Build: Disable version update script in default build
versionupdate.py causes parts of nomacs to be recompiled/relinked even
if the version did not change, which slows down development builds
considerably.

This script is only useful when the incrementor is also enabled,
therefore it is moved to that block in cmake.

The default build will always use the version defined in CMakeLists.txt,
and check that it remains in sync with version.cache file, since that is
also used by windows installer/msi builds.

Also adds cmake message with nomacs version and a warning if incrementor
is enabled since that could publish the wrong version.