fix: combine the slots of imageLoadedSignal
Because the DkViewPort::setImage may take time, the update of the "File
Info" and the metadata is better placed after that to reduce the time lag
visually. Also, the imageLoadedSignal has a `loaded` parameter that gets
unchecked by the mController updates.

Combine the three slots connected to the
DkImageLoader::imageLoadedSignal into one to make the execution ordered.
Only update the mController when the image is loaded.
Rename the slots and remove unnecessary parameters.
Also remove some unused disconnect calls.

Fixes https://github.com/nomacs/nomacs/issues/1204.