fix: do not update file info when unload image
The DkViewPort::unloadImage calls DkControlWidget::updateImage(),
but the input image container might already has the next image file path
while it is loading the file. This makes the "File Info" updates before
the new image gets displayed when switching quickly like holding on arrow
keys.

Make the call still there, but put an empty argument in to avoid this.