fix: fix DkImageStorage::compute state not reset
I did not include the `mComputeState` reset when removing the `init()`
call in 80844dd. This results in resizing the viewport do not trigger
anti-alias re-calculation.

Because when calling `compute()`, we meant to update the scaled image, so
do not check for `mComputeState == l_compute` anymore. Also move things
around and absorb the related state resetting into `compute()`.