fix: prevent manipulator preview jump back to original
Because of the imageUpdated signal triggers the preview in the
DkManipulatorWidget to update, but it calls
DkImageContainer::imageScaledToWidth, which has some cache. The preview
can jump back to original if the cache is not invalidated after actions
like rotation.

Clear the cache of the scaled images in DkImageContainer::setImage to
fix this.