fix: do not compute preview inside DkManipulatorWidget
The current implementation of the actions like rotation merges
adjacent edits of the same type, but the preview inside does not have
the undo stack info, so it can rotate with excess angle and jump back to
the correct one after the imageUpdated signal. This causes the preview
to flash or jump.

Do not calculate the preview separately to avoid this. Wait for the
whole viewport to be updated instead. This could be slower, but more
consistent.

Fixes https://github.com/nomacs/nomacs/issues/350.