fix: fix crash when rotating 8-bit image
I misunderstood the "alignment" in the QImage docs. Here it means that
the start of each row is at least 32-bit aligned, while the elements are
not necessarily so. This means we cannot treat 8-bit images as QRgb
arrays, and these images might have paddings at the end of each row.

Alter the image rotation function to handle different bit depths and the
paddings at the end of the rows.

Fixes https://github.com/nomacs/nomacs/issues/1232