Rename File: do not advance to the next file
renameFile() used to either reload the file after renaming, or advance
to the next file depending on the save file action setting. This is not
what users generally expect;and we have keyboard shortcuts for that if
the goal was to rename files efficiently.

- fix renaming a files with unsaved changes and dropping the changes;
add isEdited() to DkViewPort to support this.

- fix bug on Windows where file would be deleted if NFS is mounted case-
insensitive

Refactor renameFile()

- move into DkCentralWidget, seems like a better place for it and there
was a TODO for this.

- change "fileInfo" in messages to "file"
- add error logging
- change some wording on the dialogs
- use QNtfsPermissionCheckGuard on windows to detect more issues before
renaming