fix: incorrect plugins signal connection
Qt::UniqueConnection now gives an assertion failure with lambas on Qt6,
but only in debug builds. This makes sense because lambdas have no
metaobject so Qt cannot determine if the connection is unique or not.

Therefore, use function pointers. Also rename
things so it is more clear what slots are reserved for plugins.