Shortcuts: add global shortcut override handler
Since we now use default shortcut context, this causes a problem with
shortcut keys that conflict with widget keys. For example, when File
Explorer has input focus, "e" should jump to the first file/folder that
starts with "e" or "E" and not hide the explorer.

To resolve this, implement a shortcut override handler for the
entire application. This allows users to bind any available key
without breaking keyboard function in widgets.

Also adds DkActionEventFilter, to handle overrides for our custom
widgets that use actions. Uses this to fix the shortcuts on Notes,
which are bugged with esc-key-to-close and also do not require input
focus to trigger the shortcut.