refactor: DkTabInfo getIcon from actual image
This is part of the effort to move DkThumbNailT out of DkImageContainer.

Get the icon for the tab bar from the actual image instead of the
DkThumbNailT. This does not change the behavior because
DkImageContainerT::loadingFinished writes the loaded image to the
DkThumbNailT. Therefore we are still getting the same data.
Pass the icon size into getIcon() for caching the scaled image.