feat: single-threaded thumbnail loader
Originally, the thumbnails were loaded concurrently with DkThumbNailT.
However, since the thumbnails we load generally lives in a single
directory, i.e. one I/O device, reading the files concurrently probably
does not help.

Add DkThumbLoader to spawn a thread and accept requests for thumbnails.
The requests can be automatically queued by the QueuedConnection.
With the queue, we can prioritize reading the files that has EXIF
thumbnails and come back to decode the full image for those doesn't
later.

Place the DkThumbLoader in DkCentralWidget, the lowest common ancestor
of the widgets that use thumbnails. Rewrite DkThumbLabel so it uses
DkThumbLoader and no longer holds DkThumbNailT.

Closes https://github.com/nomacs/nomacs/issues/648