feat: detach thumbnails from image container
Thumbnails should be the preview of the files on the disk,
so they only update when the edits are written to the file.
Do not read thumbnail from the image containers in DkFilePreview.
Then we can make the DkFilePreview not dependent on DkImageContainerT.
Also gets rid of various syncing issues.

Fixes https://github.com/nomacs/nomacs/issues/663
Fixes https://github.com/nomacs/nomacs/issues/787
Closes https://github.com/nomacs/nomacs/issues/1263