fix: prevent wrong bg color at fullscreen startup
When starting nomacs fullscreen, there is a moment when the window is
first shown in default bg color before changing to fullscreen color.
This is due to the default window fill which is tied to the
QPalette::Window or perhaps the theme.

This only occurs when nomacs is started and before the viewport has been
created, and so it is disabled otherwise.