refactor: move common fade logic into a mixin
fade/visiblity logic was previously copied between three classes,
DkFadeWidget, DkFadeLabel, and DkFolderScrollBar. This was required
because these cannot have a common base due to being different QWidget
subclasses.

There is now DkFadeMixin which provides the common code. There is now a
small set of functions which must be copied so that QWidget::setVisible
is correctly overridden.

Previously, QWidget::setVisible was masked, but not overridden which
allows fade logic to be bypassed depending on the static type of the
variable. This has revealed several widgets that are now fading that
were not before. This will be addressed in a separate commit.