Fix startup glitches related to fullscreen mode (#1285)
* fix: incorrect viewport background at startup

Make base viewport background drawing available to
subclasses.

Call base viewport drawBackground() in default viewport
to fix incorrect viewport drawing before an image
is loaded.

* refactor: extract redundant code

* fix: prevent wrong bg color at fullscreen startup

When starting nomacs fullscreen, there is a moment when the window is
first shown in default bg color before changing to fullscreen color.
This is due to the default window fill which is tied to the
QPalette::Window or perhaps the theme.

This only occurs when nomacs is started and before the viewport has been
created, and so it is disabled otherwise.

* feat: smoother fullscreen transition

Prevent visual artifacts entering/exiting fullscreen mode
- ensure no painting occurs while configuring new app mode
- ensure new mode is fully configured before the switch so it won't
hitch during animation or just after animation completes.