fix: prevent redundant visibility changes in DkControlWidget
DkControlWidget::imageLoaded was called every time an image was loaded,
but it only changes behavior on the first image loaded or null image.
This adds a bunch of redundant calls to hide/show widgets when there are
no visibility changes needed.

Therefore it is changed to DkControlWidget::imagePresenceChanged which
is only called if the viewport changes between null/loaded states.