fix: DkPlayer breakage due to DkFadeMixin
DkPlayer broken as override of show() with default argument masks
DkFadeMixin::show(). Rename this to showTemporarily() to prevent that.

This matches the previous behavior except for these changes.

1) If the display mode is changed while playing, the player is not
shown. The previous behavior was to show the player and it would not
autohide unless restarting the slideshow.

2) The player controls (pause/play) cannot permanently affect the
visibility of the player. If the user toggles the player visibility
manually (P) then the change is saved.