fix: masking of function via virtual hide()
definition of virtual hide() makes a conflict with DkFadeMixin because
it also declares hide() with one default argument. It is also a conflict
with QWidget::hide() as that is not virtual.

There is no reason to mask hide()/show() in Qt, as these are not virtual
it can cause bugs. The correct method is to override setVisible(bool)
which is virtual.