fix: improper override of DkFadeMixin::setVisible()
There are some pitfalls when overriding DkFadeMixin::setVisible() due to
the recursion through QWidget::setVisible, see comment @
DkFadeMixin::setVisible.

Infinite recursion is not possible, but the function body may be
executed multiple times per call to setVisible()

Currently, none of this is strictly required but in the future it would
likely result in bugs.