perf: cancel load for obscured DkThumbLabel
The original design makes DkThumbLabel request thumbnail load when
`paint` method found no available thumbnail. This make thumbnail loading
slow when scrolling quickly in a directory with thousands of images
because we need to wait for all the thumbnails scrolled past to load
even if they are not visible anymore.

Modify the DkThumbLabel::cancelLoading() method to properly handle
cancelation of requests. Subscribe for scroll of DkThumbsView and cancel
the thumbnail requests from DkThumbLabel that are not visible.