feat(cmake): add alterative to macdeployqt for portable bundle
macdeployqt currently gives a lot of errors. It seems to make a working
bundle however it is *extremely* slow and doesn't provide a lot of
customization.

This commit adds a shell script which gives us full control over the
portable bundle and is much faster, around 10-20x. This comes from a
fundamental difference, we use the system linker to compute the
libraries rather than try to simulate what happens at run-time as in
macdeployqt.

To make this work I've also added "--about" option to nomacs to cause it
to load all libraries without a UI. Additionally, "--about" serves as a
command line version of the "About nomacs" dialog.