fix(loader): redundant loading of last open directory at startup
DkImageLoader often tries to load the last open directory, depending on
how nomacs is started. This often results in loading the dir then
throwing away the result immediately as the opened file does not match.

There doesn't seem to be a use case for this; either we open a file from
history, command line, or actions; we should not be opening anything by
default at startup.

DkImageLoader is also never constructed with a file, so remove that from
the constructor.