fix: bypass DkThumbLabel use-after-free temporarily
The loadFileSignal is connected to slots that updates the current
directory and could potentially delete the DkThumbLabel during the
process, making `mFilePath` a dangling reference.
See the discussions in https://github.com/nomacs/nomacs/pull/1262

This temporarily bypasses the possible segfault by copying the
`mFilePath`, so the value outlives the signal. We should fix the
lifetime of DkThumbLabel in the future.