cmake: disable plugins DateModified auto updating
The "DateModified" plugin metadata is written when cmake is run, and
does not contain the actual date plugin was last modified. This does
not seem to be a useful feature and also creates modified source files
every time nomacs is compiled.

If it is useful, plugin authors can update it manually when publishing
new versions.