fix: directory update results in invalid current file
Prevent folder scroll widget from loading a file if the folder size
change would give it an invalid current index.

DkImageLoader already handles directory updates/current image and this
causes a conflict.

Fixes bug when saving images to new tab which might show the wrong file
in the tab due to this effect.