feat(opendir): add setting for loading first image
Add a setting for what to do after opening a directory. The default is
to show thumbnails, this adds a setting to load the first image instead.

This behavior was mixed before (sometimes thumbs/sometimes first image)
depending on how the dir was opened (drag/drop, file menu, cmdline etc).
Now that it is consistent add this option so users can still get the old
behavior, depending on how they opened dirs.