fix: update metadata panels on image container update
Originally, the metadata panels are updated only when a new image is
loaded.
However, it is possible that the metadata changes during editing,
so these panel should update when things change even if we are still on
the same image.

This partly deals with https://github.com/nomacs/nomacs/issues/1361