refactor(page-extraction): default settings hack
Remove settings load/save in the constructor which bypasses the normal
plugin interface. This hack was used to establish the default setting so
that Batch UI would be able to see it.

The proper fix is to set the default value the first time settings are
loaded (which is performed by the plugin manager).