fix: fix mImgC of DkControlWidget become null
Some operations like cropping calls `DkViewPort::setEditedImage`, which
somehow calls `unloadImage()`. The image container inside the
mController is set to null in this function. However, because the file
is not actually changing, the `mImgC` is not set again.

These operations should not actually call `unloadImage()`, but some of
them are related to the plugins, which I am not familiar with, so
changing this could be risky. Omit the `updateImage` call if `fileChange`
is false instead for now.

Fixes https://github.com/nomacs/nomacs/issues/1360.