fix: erase xmp data with loop
Resetting rating to 0 did not work after setting rating multiple times.
For example, pressing 3, 4, and then 0 does not set the rating to 0. It
becomes 3 instead. This is due to the Xmp data are kind of "appended",
so when setting multiple times, the key could be duplicated.

Use a loop to erase all matched keys instead of the first one.