fix: copy all metadata when setting history
A bug was found when starting with an image with no rating, set the
rating to a value (e.g. 4), and "undo" cannot set the rating to 0.
When setting history, only the EXIF metadata were copied, causing that
the Xmp data cannot be cleared. So, the rating 4 can still be read from
the Xmp part.

Use the `setMetadata` method instead to copy all metadata.