fix: remove file info animation after crop
It was found that the file info panel does not show after the crop
finished. This is due to the `setVisible()` method is not synchronous,
so the `showFileInfo(true)` call is a no-op because `mFileInfoLabel`
is not visible yet.
See https://github.com/nomacs/nomacs/pull/1374.

Because we already have the crop widget fading away, the file info panel
does not need animation to have a "show up" effect. Remove this block of
code.