refactor: do not call unload image when setEditedImage
The setEditedImage calls unloadImage, which requires a fileChanged flag
to clean up some stuff. This part of the code is confusing.
Because the isPlaying is likely false during editing, move only the
related code in unloadImage into setEditedImage. Merge the two overload
of setEditedImage into one. And remove the unloadImage from the
DkBaseViewport.