fix(viewport): incorrect display with device scaling
When device scaling is enabled (deviceScaleFactorF() > 1), we need to
prevent QPainter::drawImage() from scaling. This cannot be allowed as it
breaks AA, since at 100% scale we do not map 1:1 to screen pixels.

Fix this by reporting the image size in *logical* pixels, this keeps the
transform logic unchanged. When we want the physical coordinates/sizes
we do that conversion as needed.

Additional patches are required for related features:
  - subimage extraction (rescale dialog, compression dialog)
  - paint/transform plugins
  - x/y coordinate display (eyedropper color)
  - overview panel
  - maybe printing

Fixes: #1268, #1225, #1156, #1126