fix(resize-dialog): incorrect lock button
Replace DkButton with QPushButton and stylesheet rule. DkButton seems to
be a bit convoluted and can be replaced with stylesheet rules.

The only difference for lock button seems to be how the highlighting
works; on DkButton the highlight is on the hover and colorizes the icon.
Now it looks like toolbar buttons, where the hover changes the
background color.