fix(paint-plugin): toolbars are not reshown after hiding
Overriding of setVisible() in plugin widgets has unexpected side-effects
because not only is it called by plugin manager, but also by Qt when the
widget is made visible or hidden. This makes setVisible() in widgets not
a great idea here and fixes this bug.

Remove setVisible() from widgets and move initialization/setup into the
main plugin class where it should be.

Fixes: #652