fix(viewport): incorrect image positioning/scale with AA
When AA was applied, the image may appear to squish, stretch, and/or
offset slightly vs the non-AA image. This was because the scaled image
returned by mImageStorage.image() is not the exact size called for by
the viewport transform, as that can be fractions of a pixel and we can
only scale the image to integer sizes.

To fix this, adjust the display rect so that Qt does not perform any
scaling to the image. This eliminates the squish/stretch but there will
still be a slight offset when toggling AA. This seems to be unavoidable
as Qt has to round to pixels in the end.

Also rounds the request size up to nearest pixel (was using nearest
pixel before) so that transparency pattern behind image cannot show
around the edges.