fix(translation): duplicate Portuguese/Brazil translation
The br_pt.ts and pt.ts are duplicate translation files, they have the
same language code (pt-BR) in the header and most of the strings are the
same. br_pt has many more untranslated strings, so I merged br_pt with
pt, with pt being the authority/newer file. The resulting file was
nearly identical to pt and so I am dropping br_pt.

pt is renamed to pt_BR, there will be a new nomacs_pt.ts with
traditional Portegueuse translation in a separate commit.

Fixes: #1063