fix: incorrect display scale in build info
On Wayland, the screen display scale is not the same thing as the window
scale. Display the scale we actually use in nomacs or n/a if it is
unknown. This will identify wayland compositors that do not support
fractional scaling, e.g. they will show 2.0 instead of 1.5 for example.