fix(hud): nav buttons block mouse events to hud widgets
DkHudNavigation widget - which wraps next/prev buttons - is above all
other widgets and blocks events to notes/rating when visible and not
hoving over the button. Primarily, it is not possible to click on the
star rating when next button is visible.

Remove the wrapper around nav buttons and place them in the viewport
directly. Now events are blocked only when mouse is over the nav button,
not when the nav button is visible.