feat: change edit/copy to always copy the file path
Edit/copy might copy the file path, file buffer, or nothing at all
depending on the state of the image. This is due to mixing the mime data
with the drag operation and edit/copy via createMime()

Change this to always copy the path, "Copy Buffer" remains the reliable
way to copy image pixels.

Additionally, dragging the image is more capable:
- include image buffer if file is from zip
- will always include the path if the image buffer is not available,
  rather than empty data