fix(thumbs): continuous repainting in DkThumbScene
It is possible for thumb scene to get into continous repainting (once
scroll stopped, repainting never stops). This was because pixmap cache
was too small, so thumbs that need to repaint keep getting kicked out of
the cache. Solution is to increase the pixmap cache according to number
of visible thumbnails.