fix(zoom): always-zoom-to-fit conflicts with keep zoom
Always scale to fit does not work unless "never keep zoom" is also
selected. This is counterintuitive and breaks always keep zoom with no
apparent way to resolve it.

Fix by moving always-zoom-to-fit into keep zoom enum/group so these
cannot be enabled at the same time. This is fine to do because the zoom
modes are meant to be mutually exclusive.

Fixes: #1428