fix(metadata): limit size of embedded thumbnail view
The thumbnail currently can resize the metadata panel which is visually
disruptive at best. At worst, the piece of metadata you are looking for
goes out of view.

Reduce tree view resizing due to image browsing by using a fixed height
of 160 which is the defacto standard. Some resizing is still possible if
a wide image is loaded (e.g. panorama) and the view happens to be too
narrow.