fix(metadata): undo after metadata change drops xmp/iptc metadata
If there is a metadata change like rating, rotation, or comment, using
CTRL+Z to revert the change and then saving will delete xmp and/or iptc
metadata permanently.

Fix by taking a copy of these. Any other data that is copied to the
image in saveMetaData() must also be copied in copy().

Fixes: #1401