feat: add permanent deletion for files that cannot be trashed
Trash is not supported on all volume types (especially on Windows). Add
dialog to allow these files to be deleted permanently. This is not a
remembered dialog due to destruction of files.

Change moveToTrash() to take a list of files so we can have a "yes to
all" function.

Remove most of DkThumbScene::deleteSelected(), it is not required
anymore.

Fixes: #1106